clean-arch-pt3: moved core.testing to infrastructure
