--- src/core.testing/index.test.ts
@@ -1,1 +1,0 @@
-export * from './test/utils'
